refactor: less cycles in appiled filter bars
