+++ src/components/VacancyFilterApplied/VacancyFilterApplied.tsx
@@ -4,6 +4,7 @@ import { gray7 } from '@taskany/colors';
 import { HireStream, User } from '@prisma/client';
 
 import { Group, VacancyStatus, vacancyLabels } from '../../modules/crewTypes';
+import { arrayToAppliedString } from '../../utils';
 
 import { tr } from './VacancyFilterApplied.i18n';
 
@@ -17,19 +18,6 @@ interface VacancyFilterAppliedProps {
     hrEmails?: string[];
     hiringManagers?: User[];
     hiringManagerEmails?: string[];
-}
-
-interface EntityData {
-    id: string | number;
-    name: string;
-}
-
-function mapToAppliedString(map: Record<string, EntityData>, ids: string[] | number[], title: string): string {
-    const arrayApplied = ids.reduce((acc, rec, index) => {
-        const name = index === ids.length - 1 ? map[rec].name : `${map[rec].name}, `;
-        return acc + name;
-    }, '');
-    return `${title}${arrayApplied}. `;
 }
 
 export const VacancyFilterApplied = ({
@@ -51,45 +39,19 @@ export const VacancyFilterApplied = ({
         }
 
         if (hireStreams?.length && hireStreamIds?.length) {
-            const hireStreamMap = hireStreams.reduce(
-                (acc, rec) => ({
-                    ...acc,
-                    [rec.id]: { id: String(rec.id), name: rec.name },
-                }),
-                {},
-            );
-
-            result += mapToAppliedString(hireStreamMap, hireStreamIds, tr('Streams: '));
+            result += arrayToAppliedString(hireStreams, hireStreamIds, tr('Streams: '), 'id');
         }
 
         if (hrs?.length && hrEmails?.length) {
-            const hrMap = hrs.reduce(
-                (acc, rec) => ({
-                    ...acc,
-                    [rec.email]: { name: rec.name || rec.email, id: rec.email },
-                }),
-                {},
-            );
-
-            result += mapToAppliedString(hrMap, hrEmails, tr("HR's: "));
+            result += arrayToAppliedString(hrs, hrEmails, tr("HR's: "), 'email');
         }
 
         if (hiringManagers?.length && hiringManagerEmails?.length) {
-            const hiringManagerMap = hiringManagers.reduce(
-                (acc, rec) => ({
-                    ...acc,
-                    [rec.email]: { name: rec.name || rec.email, id: rec.email },
-                }),
-                {},
-            );
-
-            result += mapToAppliedString(hiringManagerMap, hiringManagerEmails, tr('Hiring managers: '));
+            result += arrayToAppliedString(hiringManagers, hiringManagerEmails, tr('Hiring managers: '), 'email');
         }
 
         if (teams?.length && teamIds?.length) {
-            const teamMap = teams.reduce((acc, rec) => ({ ...acc, [rec.id]: rec }), {});
-
-            result += mapToAppliedString(teamMap, teamIds, tr('Teams: '));
+            result += arrayToAppliedString(teams, teamIds, tr('Teams: '), 'id');
         }
 
         return result;
